hf_space/requirements.txt
# PetPal AI Health Backend - Requirements
# Google ADK + Mem0 + FastAPI for HuggingFace Spaces

# Core FastAPI - Updated for google-adk compatibility
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-multipart>=0.0.9
pydantic>=2.5.0

# Google ADK for Multi-Agent System
google-adk>=0.5.0
google-genai>=1.0.0

# Mem0 for Persistent Memory
mem0ai>=0.1.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
